moved logs test
